implementation du formulaire Add
Implémentation du formulaire Add et son script, quelques modifs dans les autres et début du téléchargement du fichier

+++ jarditou_PHP/delete_script.php
@@ -10,50 +10,59 @@
     </p>
 </div>
 <?php
-    $pro_id = $_POST["id"];
-    //$pro_cat_id = $_POST["cat"];
-    //$pro_ref = $_POST["ref"];
-    //$pro_libelle = $_POST["lib"];
-    //$pro_description = $_POST["des"];
-    //$pro_prix = $_POST["prix"];
-    //$pro_stock = $_POST["stock"];
-    //$pro_couleur = $_POST["color"];
-    //$pro_photo = $_POST["ext"];
-
-
-    //$pro_d_ajout  = $_POST["pro_d_ajout"];
-    //$pro_d_modif = getdate();
-
-    /* if ($_POST["block"] == true)
+    if ($_SERVER["REQUEST_METHOD"] == "POST" )
     {
-        $pro_bloque = 0;
+        if (empty($_POST["id"]))
+        {
+            header("Location: liste.php");
+        }
+        else
+        {
+            $pro_id = $_POST["id"];
+            //$pro_cat_id = $_POST["cat"];
+            //$pro_ref = $_POST["ref"];
+            //$pro_libelle = $_POST["lib"];
+            //$pro_description = $_POST["des"];
+            //$pro_prix = $_POST["prix"];
+            //$pro_stock = $_POST["stock"];
+            //$pro_couleur = $_POST["color"];
+            //$pro_photo = $_POST["ext"];
+
+            //$pro_d_ajout  = $_POST["pro_d_ajout"];
+            //$pro_d_modif = getdate();
+
+            /* if ($_POST["block"] == true)
+            {
+                $pro_bloque = 0;
+            }
+            else
+            {
+                $pro_bloque = null;
+            } */
+
+            require("connexion_bdd.php"); // Inclusion de notre bibliothèque de fonctions
+
+            $db = connexionBase(); // Appel de la fonction de connexion
+            $requete = $db->prepare("DELETE FROM produits WHERE pro_id = :pro_id");
+            $requete->bindValue(":pro_id", $pro_id);
+            //$requete->bindValue(":pro_cat_id", $pro_cat_id);
+            //$requete->bindValue(":pro_ref", $pro_ref);
+            //$requete->bindValue(":pro_libelle", $pro_libelle);
+            //$requete->bindValue(":pro_description", $pro_description);
+            //$requete->bindValue(":pro_prix", $pro_prix);
+            //$requete->bindValue(":pro_stock", $pro_stock);
+            //$requete->bindValue(":pro_couleur", $pro_couleur);
+            //$requete->bindValue(":pro_photo", $pro_photo);
+            //$requete->bindValue(":pro_d_ajout", $pro_d_ajout);
+            //$requete->bindValue(":pro_d_modif", $pro_d_modif);
+            //$requete->bindValue(":pro_bloque", $pro_bloque);
+
+            $requete->execute();
+
+            $db = null;
+        }
     }
-    else
-    {
-        $pro_bloque = null;
-    } */
-
-    require("connexion_bdd.php"); // Inclusion de notre bibliothèque de fonctions
-
-    $db = connexionBase(); // Appel de la fonction de connexion
-    $requete = $db->prepare("DELETE FROM produits WHERE pro_id = :pro_id");
-    $requete->bindValue(":pro_id", $pro_id);
-    //$requete->bindValue(":pro_cat_id", $pro_cat_id);
-    //$requete->bindValue(":pro_ref", $pro_ref);
-    //$requete->bindValue(":pro_libelle", $pro_libelle);
-    //$requete->bindValue(":pro_description", $pro_description);
-    //$requete->bindValue(":pro_prix", $pro_prix);
-    //$requete->bindValue(":pro_stock", $pro_stock);
-    //$requete->bindValue(":pro_couleur", $pro_couleur);
-    //$requete->bindValue(":pro_photo", $pro_photo);
-    //$requete->bindValue(":pro_d_ajout", $pro_d_ajout);
-    //$requete->bindValue(":pro_d_modif", $pro_d_modif);
-    //$requete->bindValue(":pro_bloque", $pro_bloque);
-
-    $requete->execute();
-
-    $db = null;
-    header("Location:liste.php");
+    header("Location: liste.php");
 ?>
 <?php
     include("footer.php");
